Working on text edit reimplementation.

+++ src/Experimental/FutureKeyboard/FutureKeyboard.cpp
@@ -657,7 +657,7 @@ namespace eyegui
 		int keyOffsetY = mMode == Mode::MANY_SUGGESTION_LINES ? (int)(0.27f * mHeight) : (int)(0.4f * mHeight); // defines where keys start
 
 		// Offset for suggestions and keys to start
-		int xMargin = (mWidth - (10.f * keyWidth + 9.f * keySpace)) / 2.f;
+		int xMargin = (int) (mWidth - (10.f * keyWidth + 9.f * keySpace)) / 2.f;
 
 		// Values for suggestion line
 		int suggestionWidth = (int)((mWidth - ((2.f * xMargin) + (2.f * keySpace))) / 3.f);
@@ -672,7 +672,7 @@ namespace eyegui
 		{
 			// Move suggestion line a little bit up to improve selection by user
 			suggestionHeight = (int)(0.09f * mHeight);
-			suggestionOffsetY = (keyOffsetY - keySpace - suggestionHeight) - (mHeight * 0.0125f);
+			suggestionOffsetY = (int)((keyOffsetY - keySpace - suggestionHeight) - (mHeight * 0.0125f));
 		}
 
 		// Position empty suggestion placeholders
